refactor(application): 패키지 정리
작업내용
- 잘못 배치된 파일과 패키지 정리
