+++ src/main/java/attendance/domain/AttendanceSystem.java
@@ -4,8 +4,9 @@ import attendance.domain.checker.AttendanceChecker;
 import attendance.domain.checker.AttendanceType;
 import attendance.domain.crew.CrewStorage;
 import attendance.domain.dto.AttendanceState;
-import attendance.domain.record.AttendanceRecord;
+import attendance.domain.risk.RiskType;
 import attendance.exception.ExceptionMessage;
+import attendance.record.AttendanceRecord;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
