+++ src/main/java/attendance/domain/risk/RiskType.java
@@ -1,4 +1,4 @@
-package attendance.domain;
+package attendance.domain.risk;
 
 import java.util.List;
